requirements.txt
# Academic Review Plugin Dependencies
marker-pdf>=0.2.0
